requirements.txt
discord.py>=2.3.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
numpy>=1.21.0
aiohttp>=3.8.0